QNX/BB10: Components, UI, media, and remaining adaptations
Port remaining Chromium subsystems to QNX: compositor (cc), GPU command
buffer, UI framework, media capture/encoding, components (autofill,
sync, history, password manager, viz), and build tooling. Fixes for
QNX-specific compiler limitations, missing POSIX APIs, and platform
detection throughout the codebase.

+++ cc/base/rtree.h
@@ -207,9 +207,9 @@ void RTree<T>::Build(size_t item_count,
     // don't have enough nodes. That can happen for at most kMinChildren nodes
     // (since otherwise, we'd create a new node).
     size_t branch_count = kMaxChildren;
-    double depth = log(branches.size()) / log(branch_count);
+    double depth = log((double)branches.size()) / log((double)branch_count);
     size_t node_count =
-        static_cast<size_t>((std::pow(branch_count, depth) - 1) /
+        static_cast<size_t>((std::pow((double)branch_count, depth) - 1) /
                             (branch_count - 1)) +
         kMinChildren;
     nodes_.reserve(node_count);
